Ubacena zastita na backendu i frontu

+++ purchasesMicroservice/purchasesCore/src/main/java/com/ftn/isa/projekat/purchases/purchasesCore/adminLink/controller/AdminLinkController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -63,11 +64,14 @@ public class AdminLinkController {
 					@ApiResponse( code = 201 , message = "Created"),
 					@ApiResponse( code = 400, message= "Bad request")
 	})
-	public ResponseEntity<AdminLinkDTO> addAdminLinks(@RequestBody AdminLinkDTO dto){
-		
-		AdminLinkDTO savedAdminLink = adminLinkService.save(dto);
-		
-		return ( savedAdminLink!=null )? new ResponseEntity<AdminLinkDTO>(savedAdminLink,HttpStatus.CREATED): new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+	public ResponseEntity<AdminLinkDTO> addAdminLinks(@RequestHeader("Role") String role, @RequestBody AdminLinkDTO dto){
+		if(role.equals("ADMIN")) {
+			AdminLinkDTO savedAdminLink = adminLinkService.save(dto);
+			
+			return ( savedAdminLink!=null )? new ResponseEntity<AdminLinkDTO>(savedAdminLink,HttpStatus.CREATED): new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+			
+		}
+		return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 	}
 	
 	@DeleteMapping("/{id}")
@@ -75,11 +79,14 @@ public class AdminLinkController {
 	@ApiResponses( value = { 
 			 @ApiResponse( code = 200, message ="OK"),
 			 @ApiResponse( code = 404, message ="Not Found")})	
-	public ResponseEntity<AdminLinkDTO> deleteAdminLink(@PathVariable("id") Long id){
-		
-		AdminLinkDTO deletedAdminLinkDTO = adminLinkService.deleteById(id);
-		
-		return (deletedAdminLinkDTO.getId() != null ) ? new ResponseEntity<AdminLinkDTO>(deletedAdminLinkDTO,HttpStatus.OK) : new ResponseEntity<>(HttpStatus.NOT_FOUND);
+	public ResponseEntity<AdminLinkDTO> deleteAdminLink(@RequestHeader("Role") String role, @PathVariable("id") Long id){
+		if(role.equals("ADMIN")) {
+			AdminLinkDTO deletedAdminLinkDTO = adminLinkService.deleteById(id);
+			
+			return (deletedAdminLinkDTO.getId() != null ) ? new ResponseEntity<AdminLinkDTO>(deletedAdminLinkDTO,HttpStatus.OK) : new ResponseEntity<>(HttpStatus.NOT_FOUND);
+			
+		}
+		return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 	}
